+++ Desktop/Ludoteca/Entidad/EN_ServiciosVisita.cs
@@ -9,7 +9,7 @@ namespace Entidad
     public class EN_ServiciosVisita
     {
         public int Servicio_Id {  get; set; }
-        public int Servicio_Precio {  get; set; }
+        public double Servicio_Precio {  get; set; }
         public string ServicioName {  get; set; }
         public int  Tiempo { get; set; } //Tiempo en minutos
    }
